Fix Config Saving Issues
Due to profiles being half-implemented, the config is read from .../playmate/default/config.toml, but is written to .../playmate/config.toml.
The token cache was also being saved in the "playmate/default" profile directory, but since it should be shared across profiles, it should be saved one level up in the "playmate" directory. This commit fixes both issues.